+++ src/index.ts
@@ -18,6 +18,7 @@ export { default as GanttDependencies } from './components/GanttDependencies.vue
 export { default as GanttConflicts } from './components/GanttConflicts.vue'
 export { default as GanttSlack } from './components/GanttSlack.vue'
 export { default as GanttDeadlines } from './components/GanttDeadlines.vue'
+export { default as GanttBaselines } from './components/GanttBaselines.vue'
 export { default as GanttToday } from './components/GanttToday.vue'
 export { default as GanttZoom } from './components/GanttZoom.vue'